--- Backend/src/main/java/gwangju/ssafy/backend/domain/calendar/dto/DeleteScheduleRequest.java
@@ -1,26 +1,0 @@
-package gwangju.ssafy.backend.domain.calendar.dto;
-
-import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonFormat.Shape;
-import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
-import gwangju.ssafy.backend.domain.calendar.entity.Schedule;
-import gwangju.ssafy.backend.domain.group.entity.Group;
-import java.time.LocalDate;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-@Getter
-@Setter
-public class DeleteScheduleRequest {
-
-	private Long userId;
-	private Long scheduleId;
-
-}
