+++ Backend/src/main/java/gwangju/ssafy/backend/domain/schedule/dto/DeleteScheduleRequest.java
@@ -1,0 +1,19 @@
+package gwangju.ssafy.backend.domain.schedule.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Getter
+@Setter
+public class DeleteScheduleRequest {
+
+	private Long userId;
+	private Long scheduleId;
+
+}
